% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cercle_correlation.PLSDA}
\alias{cercle_correlation.PLSDA}
\title{Plots}
\usage{
cercle_correlation.PLSDA(object, PC1, PC2)
}
\arguments{
\item{object}{}

\item{PC1, }{first principal component}

\item{PC2, }{second principal component}
}
\value{
returns a graph which shows the projection of the individuals on 2 principal axis
}
\description{
Plots
}
\examples{
data(iris)
formule = Species~.
object = plslda.fit(formula=formule, data=iris)

cercle_correlation.PLSDA(object=object, "PC1", "PC2")
plan_factoriel.PLSDA(object=object, "PC1", "PC2")
correlationplot.PLSDA(object=object,"PC1")
propz.PLSDA(object)

}
