% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{plslda.scale}
\alias{plslda.scale}
\title{plslda.scale,
Function to standardize vectors and matrixes (center and reduce)}
\usage{
plslda.scale(X, center = TRUE, scale = TRUE)
}
\arguments{
\item{X}{the given matrix must contain numeric values}

\item{center}{set to TRUE by default, in order to center the matrix data}

\item{scale}{set to TRUE by default, in order to reduce the matrix data}
}
\value{
It returns a dataframe which is the result of the centered and reduced matrix, standardized data
}
\description{
plslda.scale,
Function to standardize vectors and matrixes (center and reduce)
}
\examples{
data(iris)
scale.t1<-plslda.scale(iris[,-5],center=TRUE, scale=TRUE)
scale.t2<-plslda.scale(iris[,-5],center=TRUE, scale=FALSE)
}
