% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sample.R
\name{plslda.split_sample}
\alias{plslda.split_sample}
\title{plslda.split_sample,
This function is used to split data into a train and a test sets}
\usage{
plslda.split_sample(formula, data, train_size = 0.7)
}
\arguments{
\item{formula, }{used to select the predictive variables}

\item{data, }{the data to split into train and test sets}

\item{train_size, }{is set to 0.7 by default (70%), which means 70% of the data goes to the train set}
}
\value{
It returns train-test splits of inputs
}
\description{
plslda.split_sample,
This function is used to split data into a train and a test sets
}
\examples{
data(iris)
formule = Species~.
split_sample.t1<-plslda.split_sample(formula=formule, data=iris)
split_sample.t2<-plslda.split_sample(formula=formule, data=iris, iris,0.5)
train1 = split_sample.t1$train
Xtrain1 = split_sample.t1$Xtrain
ytrain1 = split_sample.t1$ytrain
Xtest1 = split_sample.t1$Xtest
ytest1 = split_sample.t1$ytest


}
