% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_forward.R
\name{sel.forward}
\alias{sel.forward}
\title{sel.forward,
The function used for variables selection, following the forward method}
\usage{
sel.forward(formula, data, slentry = 0.01, verbose=FALSE)
}
\arguments{
\item{formula, }{used to select the predictive variables}

\item{data, }{dataset with many numeric variables}

\item{slentry, }{if the p value of the F test stat is lower than slentry, we stop adding variables}

\item{verbose, }{is set to FALSE by default. If set to TRUE, it prints the intermediate tables at each step}
}
\value{
It returns a dataframe with individuals and only the selected variables
}
\description{
sel.forward,
The function used for variables selection, following the forward method
}
\details{
This approach carries out a selection of variables before applying the model. It adds variables one by one, and
uses the Wilks lambda which helps determine if a variable contributes significantly through its total dispersion.
which means, in each forward step, the one variable that is added is the ones that gives the single best improvement to your model.
The idea is to keep only the variables that contribute significantly to the distance between the barycentres (class centers).
}
\examples{
data(iris)
formule = Species~
sel <- sel.forward(formula=formule, data=iris, slentry = 0.01, verbose=FALSE)


}
